refactor(artifacts): construct s3-related artifact store beans when artifact-store.type=s3
and introduce NoopArtifactStoreGetter when there's no artifact-store.type specified.
Split s3-related configuration into S3ArtifactStoreConfiguration to simplify
ArtifactStoreConfiguration and pave the way for additional artifact store implementations.

Simplify S3ArtifactStoreGetter as its S3Client is no longer optional.

Yes, this leaves artifact-store.s3.enabled in a bit of a strange place.  When it's false,
and artifact-store.type=s3, S3ArtifactStoreGetter is still enabled.  It could be that
renaming "enabled" to "enable-store" is the way forward, but there's been enough juggling
in this PR already so I'm leaving that for another time.
